react_retrieval/requirements.txt
yacs~=0.1.8
scikit-learn
timm~=0.4.12
numpy~=1.21.0
sharedmem
git+https://github.com/openai/CLIP.git
torch~=1.7.0
PyYAML~=5.4.1
Pillow~=9.0.1
torchvision~=0.8.0
tqdm~=4.62.3
protobuf~=3.20.1
faiss-cpu~=1.7.2
requests~=2.28.1
azure-storage-blob~=12.13.1
autofaiss~=2.15.3